requirements.txt
asgiref==3.8.1
crispy-bootstrap5==0.7
dj-database-url==0.5.0
Django==4.2.14
django-admin-rangefilter==0.13.0
django-allauth==0.57.2
django-crispy-forms==2.2
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
sqlparse==0.5.1
